refactor(codegen): split the instance allocation out of lower_call/new.rs (#7615)
`new.rs` was 1,988 lines against `scripts/check_file_size.sh`'s 2,000-line
cap, which blocked the Layer 1 rooting migration (#7615 slice 8) — that
migration has to ADD lines to the file, replacing `refresh_rooted_args`
and the `temp_root_scope_*` marker with a `RootedGroup`.

Pure move, no behaviour change: `lower_new_impl_inner`'s field-count
computation and its three-arm object allocation become
`new_alloc::emit_instance_alloc(ctx, class_name, class) -> String`.
`new_site_is_in_loop` moves with them (its only caller is the inline
bump-allocator arm). The boundary is a boundary rather than a cut because
none of the locals the block defines — `field_count`, `cid_str`,
`parent_cid_str`, `n_str`, `packed_keys`, `alloc_field_count` — is read
anywhere below the allocation.

No rooting decision moves with it: everything the extracted block emits
sits ABOVE the instance root, whose push is the caller's next act on the
returned handle.

new.rs 1,988 -> 1,501; new_alloc.rs 531.
