+++ crates/perry-codegen/src/lower_call/new.rs
@@ -22,7 +22,7 @@ use super::new_helpers::{
 };
 use crate::expr::{lower_expr, lower_js_args_array, nanbox_pointer_inline, temp_root, FnCtx};
 use crate::nanbox::{double_literal, POINTER_MASK_I64};
-use crate::types::{DOUBLE, I32, I64, I8, PTR};
+use crate::types::{DOUBLE, I32, I64, PTR};
 
 /// Does `new <class_name>(…)` run user code — an own or inherited constructor
 /// body, or field initializers — between the instance allocation and the value
@@ -87,32 +87,6 @@ fn reload_instance(
 
 pub(crate) use super::capture_writeback::emit_class_capture_writeback;
 use super::typed_shape_init::{emit_typed_shape_layout_declare, emit_typed_shape_layout_init};
-
-/// #7469: is the `new` site being lowered inside a **loop body**?
-///
-/// This is the gate on the inline bump allocator. Inlining removes the
-/// `js_object_alloc_class_inline_keys` FFI call and, with it, the thread-local
-/// resolutions that call performs — measured 1.81× on `churn_alloc`, 1.78× on
-/// `push_cls`. But it costs ~268 bytes of machine code **per site** (measured
-/// over 10 / 200 / 800-site programs: +0, +49,536, +214,656 bytes), so it
-/// cannot be the unconditional default against this repo's binary-size
-/// campaign — a 5,000-site application would pay ~1.3 MB.
-///
-/// Loop membership is the cheapest sound proxy for "this site runs many
-/// times", and it bounds the size cost to loop bodies. A `new` executed once
-/// keeps the outlined call and contributes nothing to binary growth.
-///
-/// `loop_targets` is reused rather than a new counter added, because it is
-/// already maintained by all three loop lowerings. It also carries `switch`
-/// frames, which push an EMPTY continue label (`switch_stmt.rs`) while every
-/// loop pushes a real one — the same discriminator `Stmt::Continue`'s
-/// scan-outward-past-switch-frames logic uses. A `new` inside a bare `switch`
-/// is therefore correctly treated as not-in-a-loop.
-fn new_site_is_in_loop(ctx: &FnCtx<'_>) -> bool {
-    ctx.loop_targets
-        .iter()
-        .any(|(continue_label, _, _)| !continue_label.is_empty())
-}
 
 /// Lower `new ClassName(args…)` — Phase C.1.
 ///
@@ -446,466 +420,9 @@ fn lower_new_impl_inner(
         arg_roots.push(slot);
     }
 
-    // Compute total field count including inherited parent fields.
-    // The runtime allocates at least 8 inline slots regardless, so this
-    // mostly matters for shapes >8 fields.
-    let mut field_count = class.fields.len() as u32;
-    // Imported classes now carry their real field_names from the source
-    // module. If the field count is still 0 (no fields info available),
-    // use a generous default as a safety net.
-    if field_count == 0 && class.constructor.is_none() {
-        field_count = 32;
-    }
-    let mut parent = class.extends_name.as_deref();
-    while let Some(parent_name) = parent {
-        if let Some(p) = ctx.classes.get(parent_name).copied() {
-            field_count += p.fields.len() as u32;
-            parent = p.extends_name.as_deref();
-        } else {
-            break;
-        }
-    }
-    // Issue #26 / #321: prefer the authoritative per-class field count computed
-    // by the source-prefix-disambiguated keys-global builder. The walk above
-    // resolves parents via `ctx.classes` — a name-keyed map that holds only
-    // ONE same-named stub — so when a cross-module parent name collides
-    // (effect's `Type` in SchemaAST.ts vs ParseResult.ts) it counts the wrong
-    // parent's fields. Using the keys-global's count keeps the allocated slot
-    // count and the header `field_count` in lockstep with the keys array,
-    // which `Object.keys()` walks. Falls back to the computed walk when this
-    // class has no keys global (anonymous / no-keys path).
-    if let Some(&authoritative) = ctx.class_field_counts.get(class_name) {
-        field_count = authoritative;
-    }
-    // #6812 (w16): a per-site empty-literal anon-shape class may carry a
-    // compile-time proven builder width. Allocate that many inline slots so
-    // the FIRST instance of the site is as wide as the runtime-learned
-    // resizes make every later one — a lone under-sized first instance
-    // permanently vetoes whole-loop clone eligibility for arrays built at
-    // the site. Capacity only: the keys array stays authoritative for
-    // enumeration, and the runtime treats header field_count as alloc_limit.
-    if class.alloc_width_hint > field_count {
-        field_count = class.alloc_width_hint;
-    }
-
-    // Allocate the object with the per-class id and (if applicable)
-    // parent class id, so the runtime registers the inheritance
-    // chain for instanceof / virtual dispatch lookups.
-    //
-    // Use `js_object_alloc_class_with_keys`, which pre-populates the
-    // `keys_array` with the class's field names in declaration order
-    // (parent fields first, walking from the deepest ancestor down,
-    // then own fields). This is REQUIRED so the LLVM PropertyGet/Set
-    // fast path's slot indices match the runtime's by-name dispatch
-    // (which walks `keys_array`). Mixing the two access patterns on
-    // the same object — e.g. constructor writes via the fast path,
-    // PropertyUpdate reads via the runtime helper — only produces
-    // consistent results when both agree on the slot mapping.
-    //
-    // The packed-keys constant is interned via the StringPool. Two
-    // classes with the same field-name set + order share one constant.
-    let cid = ctx.class_ids.get(class_name).copied().unwrap_or(0);
-    let parent_cid = class
-        .extends_name
-        .as_deref()
-        .and_then(|p| ctx.class_ids.get(p).copied())
-        .unwrap_or(0);
-    let cid_str = cid.to_string();
-    let parent_cid_str = parent_cid.to_string();
-    let n_str = field_count.to_string();
-
-    // Fast path: if the class has a per-class keys global (built once
-    // at module init via `js_build_class_keys_array`), emit INLINE
-    // bump-allocator IR — no function call into the runtime at all on
-    // the hot path. The runtime exposes a `InlineArenaState` struct
-    // (data ptr at offset 0, current bump offset at offset 8, current
-    // block size at offset 16) via `js_inline_arena_state()`. We call
-    // that ONCE per JS function entry (cached in `arena_state_slot`)
-    // and then emit a 5-instruction bump check + GcHeader/ObjectHeader
-    // store sequence at every `new ClassName()` site. The slow path
-    // (block overflow) calls `js_inline_arena_slow_alloc` which syncs
-    // the inline state back to the underlying arena, allocates a new
-    // block, and updates the inline state.
-    //
-    // Cycles per inlined alloc on the M-series fast path:
-    //    load offset       (1)
-    //    add+and align     (2)
-    //    add new_offset    (1)
-    //    load size + cmp   (2)
-    //    cond br           (predicted, 0)
-    //    store offset      (1)
-    //    load data + gep   (2)
-    //    write GcHeader    (1)  — packed i64 store
-    //    write ObjectHeader×2 (2) — packed i64 stores
-    //    write keys_ptr    (1)
-    //  total: ~13 cycles vs ~140 cycles for the function-call path.
-    //
-    // Layout assumption: GcHeader is 8 bytes
-    //    {obj_type:u8, gc_flags:u8, _reserved:u16, size:u32}
-    // and ObjectHeader is 24 bytes
-    //    {object_type:u32, class_id:u32, parent_class_id:u32,
-    //     field_count:u32, keys_array:*ptr}
-    // followed by `max(field_count, 8)` 8-byte field slots. The user
-    // pointer the rest of the codegen sees is `raw + 8` (i.e. the
-    // ObjectHeader address) — same as what
-    // `js_object_alloc_class_inline_keys` returns.
-    //
-    // Layout constants are duplicated here from the runtime; if
-    // `GcHeader` or `ObjectHeader` ever change in
-    // `crates/perry-runtime/src/{gc,object}.rs`, update both sides.
-    let obj_handle = if class.extends_expr.is_some() {
-        // Wall 45: dynamic-parent subclass (`class X extends _mod.default`).
-        // The parent's field layout is unknown at this compile time (the
-        // `extends` target is an unresolvable cross-module value, so the
-        // parent-chain walk above contributed 0 fields and `field_count` /
-        // `packed_keys` cover only X's OWN fields). Allocating with that
-        // own-only layout under-sizes and mis-lays-out the instance: the
-        // parent's constructor and inherited methods address the inherited
-        // fields at the PARENT's slot indices (parent fields first), which fall
-        // past X's own slots → OOB heap reads (captures read as garbage).
-        // Route to `js_object_alloc_class_dynamic_parent`, which resolves the
-        // runtime-registered parent edge + keys-array (both established at
-        // module init by `js_register_class_parent_dynamic` /
-        // `js_build_class_keys_array`, before any `new X()`) and allocates with
-        // the merged `[parent keys..] ++ [own keys..]` layout. Bypasses the
-        // inline bump-alloc fast path (which would bake the wrong layout).
-        let mut packed_keys = String::new();
-        for f in &class.fields {
-            if f.key_expr.is_some() {
-                continue;
-            }
-            packed_keys.push_str(&f.name);
-            packed_keys.push('\0');
-        }
-        let keys_idx = ctx.strings.intern(&packed_keys);
-        let keys_entry = ctx.strings.entry(keys_idx);
-        let keys_global = format!("@{}", keys_entry.bytes_global);
-        let keys_len_str = keys_entry.byte_len.to_string();
-        ctx.block().call(
-            I64,
-            "js_object_alloc_class_dynamic_parent",
-            &[
-                (I32, &cid_str),
-                (I32, &n_str),
-                (PTR, &keys_global),
-                (I32, &keys_len_str),
-            ],
-        )
-    } else if let Some(keys_global_name) = ctx.class_keys_globals.get(class_name).cloned() {
-        // [#bloat] Outline the per-`new`-site allocator EXCEPT inside a loop.
-        //
-        // Outlining collapses ~145 lines of per-class-constant IR per site into
-        // a single `js_object_alloc_class_inline_keys` call (~3 lines), which
-        // performs the identical bump alloc + header init + slot zero-fill and
-        // returns the same user pointer. That is the right default for code
-        // size, and the size half of the original measurement still holds:
-        // ~268 bytes of machine code per site, +214,656 bytes over an 800-site
-        // program.
-        //
-        // The SPEED half of that measurement has since inverted. It read
-        // "~17% faster on an 8M-allocation loop (the inline bump bloated the
-        // hot loop, hurting icache/regalloc more than the saved call)"; today
-        // the outlined form is **1.81× SLOWER** on `churn_alloc` and 1.78× on
-        // `push_cls`. Nothing about the inline bump changed — everything
-        // *around* the allocation got cheaper (#7474, #7486, #7487, #7501,
-        // #7525, #7532, #7535, #7536, #7552), so the surviving FFI call and
-        // the thread-local resolutions it performs now dominate what the
-        // inline bump's code bloat costs. On Darwin those resolutions cannot
-        // be made cheaper — Mach-O has no local-exec TLS model, and building
-        // the runtime with `-Ztls-model=local-exec` leaves the `blr` through
-        // the TLV descriptor byte-identical (measured: 1.02×). Only their
-        // COUNT can be reduced, and inlining removes them outright.
-        //
-        // So the choice is per site, not global: a `new` inside a loop takes
-        // the inline bump (it runs many times, and the size cost is bounded to
-        // loop bodies); everything else keeps the outlined call and
-        // contributes nothing to binary growth. `PERRY_INLINE_NEW=1` still
-        // forces the inline form everywhere, for A/B measurement.
-        //
-        // NOTE the env test is `is_none()`: `PERRY_INLINE_NEW=""` *enables*
-        // the inline path, because an empty string is `Some("")`.
-        let force_inline_new = std::env::var_os("PERRY_INLINE_NEW").is_some();
-        if !force_inline_new && !new_site_is_in_loop(ctx) {
-            let keys_slot = if let Some(s) = ctx.class_keys_slots.get(class_name).cloned() {
-                s
-            } else {
-                let s = ctx.func.entry_init_load_global(&keys_global_name, I64);
-                ctx.class_keys_slots
-                    .insert(class_name.to_string(), s.clone());
-                s
-            };
-            let keys_ptr = ctx.block().load(I64, &keys_slot);
-            ctx.pending_declares.push((
-                "js_object_alloc_class_inline_keys".to_string(),
-                I64,
-                vec![I32, I32, I32, I64],
-            ));
-            ctx.block().call(
-                I64,
-                "js_object_alloc_class_inline_keys",
-                &[
-                    (I32, &cid_str),
-                    (I32, &parent_cid_str),
-                    (I32, &field_count.to_string()),
-                    (I64, &keys_ptr),
-                ],
-            )
-        } else {
-            // Compile-time layout constants.
-            const GC_HEADER_SIZE: u64 = 8;
-            // arm64_32 watchOS: `size_of::<ObjectHeader>()` is 24 on 64-bit but
-            // 20 on ILP32 (4-byte `keys_array` pointer). Derive from the target
-            // triple so the inline alloc size and field-region base match the
-            // target-compiled runtime (no-op on 64-bit; see `target_layout`).
-            let object_header_size: u64 =
-                crate::target_layout::object_header_size_bytes(ctx.target_triple);
-            // #6759 Phase B: pointer width for the trailing `meta` header
-            // field (computed here, before `ctx.block()` mutably borrows).
-            let meta_ptr_size: u64 = if crate::target_layout::target_is_ilp32(ctx.target_triple) {
-                4
-            } else {
-                8
-            };
-            const FIELD_SLOT_SIZE: u64 = 8;
-            // Inline-slot floor — MUST match perry-runtime `object::INLINE_SLOT_FLOOR`
-            // (they independently pad `new` objects to the same minimum; a mismatch
-            // where codegen allocs fewer slots than the runtime's get/set bound-check
-            // assumes is heap corruption). Lowered 8->4 to shrink small-object footprint.
-            const MIN_FIELD_SLOTS: u64 = 4;
-            const GC_TYPE_OBJECT: u64 = 2;
-            const GC_FLAG_ARENA: u64 = 0x02;
-            // PR #1146: pointer-free hint for inline-allocated regular
-            // objects. The field-store sites issue per-slot
-            // `js_gc_note_slot_layout` so the GC sees real pointer-bearing
-            // slots regardless of this initial tag.
-            const GC_LAYOUT_POINTER_FREE: u64 = 0x4000;
-            const OBJECT_TYPE_REGULAR: u64 = 1;
-
-            let alloc_field_count = std::cmp::max(field_count as u64, MIN_FIELD_SLOTS);
-            let payload_size = object_header_size + alloc_field_count * FIELD_SLOT_SIZE;
-            // Round the whole allocation up to FIELD_SLOT_SIZE (8). The inline
-            // bump allocator's offset invariant (below) requires every
-            // allocation to be a multiple of 8; on ILP32 `object_header_size`
-            // is 20, so an unpadded total is 4-skewed (e.g. 92) and would
-            // misalign the next bump. No-op on 64-bit (8 + 24 + 8·n is already
-            // 8-aligned → 96 for ≤8 fields).
-            let total_size = (GC_HEADER_SIZE + payload_size).next_multiple_of(FIELD_SLOT_SIZE);
-            let total_size_str = total_size.to_string();
-
-            // Lazy: allocate the per-function arena-state slot on the
-            // first `new` we see. The slot init (`call @js_inline_arena_state`
-            // + store) lives in the entry block via `entry_init_call_ptr`,
-            // so it dominates every reachable use.
-            let arena_state_slot = if let Some(slot) = ctx.arena_state_slot.clone() {
-                slot
-            } else {
-                let slot = ctx.func.entry_init_call_ptr("js_inline_arena_state");
-                ctx.arena_state_slot = Some(slot.clone());
-                slot
-            };
-
-            // Hoist the per-class `keys_array` global load to the function
-            // entry block (cached in a stack slot per class). Without this
-            // hoisting, LLVM would reload `@perry_class_keys_<class>` on
-            // every loop iteration, because the loop body's `call
-            // @js_inline_arena_slow_alloc` blocks LICM — LLVM can't prove
-            // the call doesn't modify the global.
-            let keys_slot = if let Some(s) = ctx.class_keys_slots.get(class_name).cloned() {
-                s
-            } else {
-                let s = ctx.func.entry_init_load_global(&keys_global_name, I64);
-                ctx.class_keys_slots
-                    .insert(class_name.to_string(), s.clone());
-                s
-            };
-            let keys_ptr = ctx.block().load(I64, &keys_slot);
-
-            // Inline bump-allocator IR.
-            let blk = ctx.block();
-            let state_ptr = blk.load(PTR, &arena_state_slot);
-
-            // offset = state.offset (at byte offset 8 in InlineArenaState).
-            // The offset is invariant 8-aligned: arena blocks start at offset 0
-            // (8-aligned), every allocation is a multiple of 8 (`total_size`
-            // includes the 8-byte GcHeader and `MIN_FIELD_SLOTS=4` slots ×
-            // 8 bytes), and `js_inline_arena_slow_alloc` only ever swings the
-            // state to `block.offset` which is also always 8-aligned. So we
-            // skip the `(offset + 7) & -8` align-up step entirely — saves
-            // 2 instructions per iter on the hot path.
-            let offset_field_ptr = blk.gep(I8, &state_ptr, &[(I64, "8")]);
-            let offset_val = blk.load(I64, &offset_field_ptr);
-            let aligned_off = offset_val.clone();
-
-            // new_offset = aligned + total_size
-            let new_offset = blk.add(I64, &aligned_off, &total_size_str);
-
-            // size = state.size (at byte offset 16)
-            let size_field_ptr = blk.gep(I8, &state_ptr, &[(I64, "16")]);
-            let size_val = blk.load(I64, &size_field_ptr);
-
-            // fits = new_offset <= size
-            let fits = blk.icmp_ule(I64, &new_offset, &size_val);
-
-            // Set up fast/slow/merge basic blocks.
-            let fast_idx = ctx.new_block("alloc.fast");
-            let slow_idx = ctx.new_block("alloc.slow");
-            let merge_idx = ctx.new_block("alloc.merge");
-            let fast_label = ctx.block_label(fast_idx);
-            let slow_label = ctx.block_label(slow_idx);
-            let merge_label = ctx.block_label(merge_idx);
-
-            ctx.block().cond_br(&fits, &fast_label, &slow_label);
-
-            // ---- Fast path: bump and return data + aligned ----
-            ctx.current_block = fast_idx;
-            let blk = ctx.block();
-            // GC_STORE_AUDIT(INIT): inline arena bump offset is allocator metadata, not a JS heap edge.
-            blk.store(I64, &new_offset, &offset_field_ptr);
-            // data ptr is at byte offset 0 in InlineArenaState
-            let data_ptr = blk.load(PTR, &state_ptr);
-            let raw_fast = blk.gep(I8, &data_ptr, &[(I64, &aligned_off)]);
-            let fast_pred_label = blk.label.clone();
-            blk.br(&merge_label);
-
-            // ---- Slow path: call into the runtime ----
-            ctx.current_block = slow_idx;
-            let raw_slow = ctx.block().call(
-                PTR,
-                "js_inline_arena_slow_alloc",
-                &[(PTR, &state_ptr), (I64, &total_size_str), (I64, "8")],
-            );
-            let slow_pred_label = ctx.block().label.clone();
-            ctx.block().br(&merge_label);
-
-            // ---- Merge: phi the raw pointer, write headers, NaN-box ----
-            ctx.current_block = merge_idx;
-            let blk = ctx.block();
-            let raw = blk.phi(
-                PTR,
-                &[(&raw_fast, &fast_pred_label), (&raw_slow, &slow_pred_label)],
-            );
-
-            // Write GcHeader (8 bytes) as a single i64 store. Field
-            // packing (little-endian):
-            //   bits  0..7   = obj_type (u8)
-            //   bits  8..15  = gc_flags (u8)
-            //   bits 16..31  = _reserved (u16)
-            //   bits 32..63  = size (u32)
-            let gc_packed: u64 = GC_TYPE_OBJECT
-                | (GC_FLAG_ARENA << 8)
-                | (GC_LAYOUT_POINTER_FREE << 16)
-                | ((total_size as u64) << 32);
-            // GC_STORE_AUDIT(INIT): inline headers initialize freshly allocated unpublished object storage.
-            blk.store(I64, &gc_packed.to_string(), &raw);
-
-            // Write ObjectHeader at raw + 8.
-            // First 8 bytes: object_type (u32, low) | class_id (u32, high)
-            let oh_addr_1 = blk.gep(I8, &raw, &[(I64, "8")]);
-            let oh_word_1: u64 = OBJECT_TYPE_REGULAR | ((cid as u64) << 32);
-            blk.store(I64, &oh_word_1.to_string(), &oh_addr_1);
-
-            // Second 8 bytes: parent_class_id (u32, low) | field_count (u32, high)
-            let oh_addr_2 = blk.gep(I8, &raw, &[(I64, "16")]);
-            let oh_word_2: u64 = (parent_cid as u64) | ((field_count as u64) << 32);
-            blk.store(I64, &oh_word_2.to_string(), &oh_addr_2);
-
-            // Third 8 bytes: keys_array pointer. The keys_ptr we loaded
-            // above is an i64 (carries the ArrayHeader address); store as
-            // i64 since the underlying memory is 8 bytes either way.
-            let oh_addr_3 = blk.gep(I8, &raw, &[(I64, "24")]);
-            // GC_STORE_AUDIT(INIT): keys_array edge is installed before publishing the new object.
-            blk.store(I64, &keys_ptr, &oh_addr_3);
-
-            // #6759 Phase B: null the `meta` record pointer — the LAST header
-            // field, at header offset (object_header_size - pointer_size).
-            // Pointer-width store: on ILP32 the field is 4 bytes at a
-            // 4-aligned offset, and an i64 store there would violate the
-            // arm64_32 `i64:64` ABI alignment (and spill into slot 0).
-            let meta_off = GC_HEADER_SIZE + object_header_size - meta_ptr_size;
-            let meta_addr = blk.gep(I8, &raw, &[(I64, &meta_off.to_string())]);
-            // GC_STORE_AUDIT(INIT): fresh inline object starts with no per-object meta record (#6759 B).
-            let meta_store_ty = if meta_ptr_size == 4 { I32 } else { I64 };
-            blk.store(meta_store_ty, "0", &meta_addr);
-
-            // PerryTS/perry#4717: zero-fill the field slots with `undefined`, mirroring
-            // `js_object_alloc_with_parent` (runtime object/alloc.rs), which deliberately
-            // initializes ALL `max(field_count, 8)` slots "to prevent stale data from
-            // previously freed GC objects from bleeding through." This inline bump path
-            // wrote only the headers and left the slots uninitialized, so a field
-            // read-before-write — or a GC that scans the still-constructing instance —
-            // observed stale arena bytes. When those bytes were a previously-freed
-            // `undefined`/pointer (e.g. `marked`'s `this.defaults`), the constructor
-            // crashed with "Cannot read properties of undefined". Slots start at
-            // raw + GcHeader(8) + ObjectHeader(24) = raw + 32.
-            for i in 0..alloc_field_count {
-                let slot_off = GC_HEADER_SIZE + object_header_size + i * FIELD_SLOT_SIZE;
-                let slot_ptr = blk.gep(I8, &raw, &[(I64, &slot_off.to_string())]);
-                // GC_STORE_AUDIT(INIT): freshly allocated inline object slot initialized to undefined.
-                blk.store(I64, crate::nanbox::TAG_UNDEFINED_I64, &slot_ptr);
-            }
-
-            // User pointer = raw + 8 (the ObjectHeader address — what the
-            // function-call path returned). Convert to i64 to match what
-            // the existing nanbox_pointer_inline expects.
-            let user_ptr = blk.gep(I8, &raw, &[(I64, "8")]);
-            blk.ptrtoint(&user_ptr, I64)
-        }
-    } else {
-        // Fallback: build the packed-keys string at this site and
-        // call the slower SHAPE_CACHE-aware allocator. Used when the
-        // class isn't in `class_keys_globals` (e.g. anonymous /
-        // synthetic classes that compile_module doesn't pre-emit a
-        // global for).
-        let mut packed_keys = String::new();
-        let mut parent_chain: Vec<&perry_hir::Class> = Vec::new();
-        let mut p = class.extends_name.as_deref();
-        while let Some(parent_name) = p {
-            if let Some(pc) = ctx.classes.get(parent_name).copied() {
-                parent_chain.push(pc);
-                p = pc.extends_name.as_deref();
-            } else {
-                break;
-            }
-        }
-        // Skip computed-key fields: their key is an expression evaluated at
-        // construction time, not a stable string, so they don't get an inline
-        // slot. The runtime stores them via IndexSet → js_object_set_field /
-        // js_object_set_symbol_property paths in `apply_field_initializers_recursive`.
-        // Including their synthetic `__computed_field_*` names in packed_keys
-        // would surface them as enumerable own properties on Object.keys().
-        for pc in parent_chain.iter().rev() {
-            for f in &pc.fields {
-                if f.key_expr.is_some() {
-                    continue;
-                }
-                packed_keys.push_str(&f.name);
-                packed_keys.push('\0');
-            }
-        }
-        for f in &class.fields {
-            if f.key_expr.is_some() {
-                continue;
-            }
-            packed_keys.push_str(&f.name);
-            packed_keys.push('\0');
-        }
-        let keys_idx = ctx.strings.intern(&packed_keys);
-        let keys_entry = ctx.strings.entry(keys_idx);
-        let keys_global = format!("@{}", keys_entry.bytes_global);
-        let keys_len_str = keys_entry.byte_len.to_string();
-
-        ctx.block().call(
-            I64,
-            "js_object_alloc_class_with_keys",
-            &[
-                (I32, &cid_str),
-                (I32, &parent_cid_str),
-                (I32, &n_str),
-                (PTR, &keys_global),
-                (I32, &keys_len_str),
-            ],
-        )
-    };
+    // #7615 slice 8: the field-count computation and the three-arm instance
+    // allocation moved verbatim to `new_alloc.rs` (see its header for why).
+    let obj_handle = super::new_alloc::emit_instance_alloc(ctx, class_name, class);
     // #7154: root the instance for the duration of the constructor body.
     //
     // Until now the instance existed ONLY as an SSA register while that body
